feat: added wallet / leaderboard in the tutorial section
