+++ client/src/lib/components/tutorial/ui.svelte
@@ -1,0 +1,34 @@
+<script lang="ts">
+  import { uiStore } from '$lib/stores/ui.store.svelte';
+  import AuctionModal from '$lib/components/auction/auction-modal.svelte';
+  import BuyModal from '$lib/components/buy/buy-modal.svelte';
+  import LandHud from '$lib/components/land/hud/land-hud.svelte';
+  import LandInfoModal from '$lib/components/land/land-info-modal.svelte';
+  import Toolbar from '$lib/components/toolbar/toolbar.svelte';
+  import WalletLookup from './wallet.svelte';
+</script>
+
+<div class="z-50 absolute top-0 left-0">
+  <!-- Wallet Lookup -->
+  <WalletLookup />
+
+  <!-- Leaderboard-->
+  <!-- <Leaderboard /> -->
+
+  <!-- Tile HUD -->
+  <!-- <LandHud />
+
+  <Toolbar /> -->
+
+  <!-- Modals -->
+  <!-- {#if uiStore.showModal}
+    {#if uiStore.modalType == 'buy'}
+      <BuyModal />
+    {:else if uiStore.modalType == 'bid'}
+      <AuctionModal />
+    {:else if uiStore.modalType == 'land-info'}
+      <LandInfoModal />
+    {/if}
+  {/if}
+   -->
+</div>
